Refactor clib back into a single class
Simplifies the code and removes duplication of docstrings.
