+++ gmt/clib/__init__.py
@@ -1,4 +1,4 @@
 """
 Low-level wrappers for the GMT C API using ctypes
 """
-from .context_manager import LibGMT
+from .core import LibGMT
